+++ Necromancy.Server/Chat/Command/Commands/OnHit.cs
@@ -1,0 +1,25 @@
+using Arrowgene.Services.Buffers;
+using Necromancy.Server.Common;
+using Necromancy.Server.Model;
+using Necromancy.Server.Packet.Id;
+
+namespace Necromancy.Server.Chat.Command.Commands
+{
+    //  battle report attack on hit. 
+    public class OnHit : ServerChatCommand
+    {
+        public OnHit(NecServer server) : base(server)
+        {
+        }
+
+        public override void Execute(string[] command, NecClient client, ChatMessage message, ChatResponse response)
+        {
+            IBuffer res = BufferProvider.Provide();
+            Router.Send(client, (ushort) AreaPacketId.recv_battle_report_action_attack_onhit, res,
+                ServerType.Area);
+        }
+
+        public override AccountStateType AccountState => AccountStateType.User;
+        public override string Key => "OnHit";
+    }
+}
