+++ Necromancy.Server/Chat/Command/Commands/SendEventEnd.cs
@@ -1,0 +1,25 @@
+using Arrowgene.Services.Buffers;
+using Necromancy.Server.Common;
+using Necromancy.Server.Model;
+using Necromancy.Server.Packet.Id;
+
+namespace Necromancy.Server.Chat.Command.Commands
+{
+    //failsafe to end events when frozen
+    public class SendEventEnd : ServerChatCommand
+    {
+        public SendEventEnd(NecServer server) : base(server)
+        {
+        }
+
+        public override void Execute(string[] command, NecClient client, ChatMessage message, ChatResponse response)
+        {
+            IBuffer res = BufferProvider.Provide();
+            res.WriteByte(0);
+            Router.Send(client, (ushort) AreaPacketId.recv_event_end, res, ServerType.Area);
+        }
+
+        public override AccountStateType AccountState => AccountStateType.User;
+        public override string Key => "EndEvent";
+    }
+}
