+++ Necromancy.Server/Chat/Command/Commands/SendMapEntry.cs
@@ -1,0 +1,32 @@
+using System;
+using Arrowgene.Services.Buffers;
+using Necromancy.Server.Common;
+using Necromancy.Server.Model;
+using Necromancy.Server.Packet.Id;
+
+namespace Necromancy.Server.Chat.Command.Commands
+{
+    public class SendMapEntry : ServerChatCommand
+    {
+        public SendMapEntry(NecServer server) : base(server)
+        {
+        }
+
+        public override void Execute(string[] command, NecClient client, ChatMessage message, ChatResponse response)
+        {
+            int mapId = Convert.ToInt32(command[0]);
+
+
+            Map map = Server.Map.Get(mapId);
+
+            //map.Enter(client);
+
+            IBuffer res = BufferProvider.Provide();
+            res.WriteInt32(0);
+            Router.Send(client, (ushort) AreaPacketId.recv_map_entry_r, res, ServerType.Area);
+        }
+
+        public override AccountStateType AccountState => AccountStateType.User;
+        public override string Key => "MapEntry";
+    }
+}
